MetricEventSource improve error reporting
When an exception occured on the collector thread we were previously printing
message and stack trace. However if the exception was caused by an inner
exception the root cause information was lost. Recording Exception.ToString()
instead of Exception.Message and Exception.StackTrace gives a more informative
error message that includes InnerException details.