WIP
This change intends to mitigate a debugger deadlock issue by emulating the behavior users would get
if they had disabled tiered compilation and rejit on attach when launching under the debugger.
The code looks like it should do the right thing but at this point it is completely untested, both that
it disables the features as it should, and that it doesn't cause regressions in the debugging experience.

Initially when I talked with Kount I was thinking of doing something that would also attempt to
handle debugger attach, but given that we don't even have someone ready to test it I decided
to be as consverative as possible about causing regressions. Given that this change should do all of
its work at startup means we should get two deterministic and easily testable behaviors, one where
the app is debugged at startup and the other where it is not.

I'm also not certain if tactics will want to take this change so a bar check could be useful. Unlike a
usual servicing issue we know that customers have hit it in the past and likely more customers are
hitting it without realizing it, but we don't have someone with a repro waiting to test it. Either we'd
be submitting it speculatively or we could sit on the fix and wait until we got another complaint.